@@ -37,12 +37,35 @@ class Unit extends Number {
 
     public static function doRender(object $data, array $options = []) {
 
-        if ($data->value == 0) {
+        /**
+         * @see https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Types#quantities
+         */
+        if ($data->value == 0 && !in_array(strtolower($data->unit), ['s', 'ms', 'hz', 'khz', 'dpi', 'dpcm', 'dppx', 'x'])) {
 
             return '0'.(isset($options['omit_unit']) && isset($data->unit) && $options['omit_unit'] == false ? $data->unit : '');
         }
 
         $unit = !empty($options['omit_unit']) && $options['omit_unit'] == $data->unit ? '' : $data->unit;
+
+        if ($data->value == 0) {
+
+            $unit = strtolower($data->unit);
+
+            if ($unit == 'ms') {
+
+                $unit = 's';
+            }
+
+            else if ($unit == 'khz') {
+
+                $unit = 'hz';
+            }
+        }
+
+        if ($unit == 'dppx') {
+
+            $unit = 'x';
+        }
 
         if (!empty($options['compress'])) {
 
